% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_PoisMLE.R
\docType{class}
\name{PoisMLE-class}
\alias{PoisMLE-class}
\alias{PoissonMLE}
\title{Define the Poisson MLE Class}
\description{
An Object of class \code{PoisMLE} can be created by the \code{mlePois()} function.
}
\details{
An Object of the class `PoisMLE` has the following slots:
\itemize{
\item \code{y} The original data, which is count data.
\item \code{MLE} The maximum likelihood estimator (MLE) for this data.
\item \code{LL} The natural logarithm of the MLE.
\item \code{SE} The standard error for the MLE.
\item \code{SEtype} The method used to calculate the standard error.
}
}
\seealso{
\code{\link{logLikelihood}}, \code{\link{mle}}, \code{\link{standardError}}, \code{\link{estimatePoisson}}
}
\author{
Jordan Duffin Wong: \email{jordan.d.wong@wustl.edu}
}
