% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePois.R
\name{estimatePois}
\alias{estimatePois}
\alias{estimatePoisson}
\title{Estimate Poisson}
\usage{
estimatePois(y, SEtype, B = 1000)
}
\arguments{
\item{y}{The vector of observed data}

\item{SEtype}{The type of SE, can choose either `basic` or `bootstrap`}

\item{B}{The number of bootstrap iterations to run if `SEtype` is `bootstrap`. Defaults to 1000.}
}
\value{
A single numeric output:
    \item{PoisMLE}{An object of class `PoisMLE` which contains the MLE, Log-Likelihood, Standard Error, and type of Standard Error used.}
}
\description{
Estimates the Poisson Distribution of a vector of data and returns an object of class `PoisMLE`.
}
\examples{
set.seed(666)
y <- rpois(1000, 10)
estimatePois(y, "basic")
estimatePois(y, "bootstrap", B = 1000)

}
\seealso{
\code{\link{logLik}}, \code{\link{mle}}, \code{\link{standardError}}
}
\author{
Jordan Duffin Wong: \email{jordan.d.wong@wustl.edu}
}
