% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\alias{logLikelihood}
\title{Log-Likelihood}
\usage{
logLik(y, lambda)
}
\arguments{
\item{y}{The vector of observed data}

\item{lambda}{the assumed value of \eqn{\lambda} lambda}
}
\value{
A single numeric output:
    \item{log_likelihood}{The log-likelihood for the observed data conditioning on assumed value of lambda.}
}
\description{
Calculates the Log-Likelihood for observed data
}
\examples{
set.seed(666)
y <- sample(x = 1:50, size = 25, replace = TRUE)
logLik(y, lambda = 3)
}
\seealso{
\code{\link{mle}}, \code{\link{standardError}}, \code{\link{estimatePoisson}}
}
\author{
Jordan Duffin Wong: \email{jordan.d.wong@wustl.edu}
}
