% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Maximum Likelihood Estimation}
\usage{
mle(y)
}
\arguments{
\item{y}{The vector of observed data}
}
\value{
A single numeric output:
    \item{MLE}{The MLE for the observed data.}
}
\description{
Calculates the Maximum Likelihood Estimator (MLE) for observed data
}
\examples{
set.seed(666)
y <- sample(x = 1:50, size = 25, replace = TRUE)
mle(y)
}
\seealso{
\code{\link[PoisMLE]{logLik}}, \code{\link[PoisMLE]{standarderror}}, \code{\link[PoisMLE]{estimatePoisson}}
}
\author{
Jordan Duffin Wong: \email{jordan.d.wong@wustl.edu}
}
